Altered code to remove runtime warnings

+++ RazorDBTests/ManifestTests.cs
@@ -215,26 +215,23 @@ namespace RazorDBTests {
         public void TestManifestFileRollover() {
 
             var path = Path.GetFullPath("TestData\\TestManifestFileRollover");
-            if (!Directory.Exists(path))
-                Directory.CreateDirectory(path);
+            if (!Directory.Exists(path)) Directory.CreateDirectory(path);
 
             // Remove the file if it exists
             var filename = Config.ManifestFile(path);
-            if (File.Exists(filename))
-                File.Delete(filename);
+            if (File.Exists(filename)) File.Delete(filename);
 
             var mf = new Manifest(path);
-            for (int i = 0; i < Config.ManifestVersionCount - 10; i++) {
-                var level = mf.NextVersion(1);
-            }
+			var manifestSize = new FileInfo(filename).Length;
+			var newManifestSize = new FileInfo(filename).Length;
 
-            var manifestSize = new FileInfo(filename).Length;
+            for (int i = 0; i < Config.ManifestVersionCount - 10; i++) {
+                mf.NextVersion(1);
+            }
 
             for (int i = 0; i < 100; i++) {
-                var level = mf.NextVersion(1);
+                mf.NextVersion(1);
             }
-
-            var newManifestSize = new FileInfo(filename).Length;
 
             // The new file should be smaller than the previous one, because it should have rolled over.
             Assert.Less(newManifestSize, manifestSize);
